# api/openapi.yaml
openapi: 3.0.3
info:
  title: host-rutebayar
  version: 0.1.0
  description: Orchestrator API draft for marketplace-host payment routing.
servers:
  - url: /v1
components:
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-Host-Secret
  schemas:
    CreatePaymentRequest:
      type: object
      properties:
        host_id:
          type: string
        product_id:
          type: string
        buyer_ref:
          type: string
        env:
          type: string
      required:
        - host_id
        - product_id
    CreatePaymentResponse:
      type: object
      properties:
        reference:
          type: string
        status:
          type: string
        checkout_url:
          type: string
      required:
        - reference
        - status
    PaymentStatus:
      type: object
      properties:
        reference:
          type: string
        status:
          type: string
        host_id:
          type: string
        product_id:
          type: string
        checkout_url:
          type: string
        host_fee_amount:
          type: integer
        gross_amount:
          type: integer
        net_amount:
          type: integer
      required:
        - reference
        - status
    WebhookPayload:
      type: object
      properties:
        reference:
          type: string
        status:
          type: string
        idempotency_key:
          type: string
      required:
        - reference
        - status
        - idempotency_key
    WebhookResponse:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    FeePolicy:
      type: object
      properties:
        type:
          type: string
        value:
          type: number
        currency:
          type: string
        rounding:
          type: string
        min_fee:
          type: integer
          nullable: true
        max_fee:
          type: integer
          nullable: true
        policy_version:
          type: string
      required:
        - type
        - value
        - currency
    RegisterHostRequest:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        callback_urls:
          type: array
          items:
            type: string
        callback_allowlist:
          type: array
          items:
            type: string
        notification_key:
          type: string
        host_secret:
          type: string
        webhook_secret:
          type: string
      required:
        - id
        - name
        - notification_key
        - host_secret
        - webhook_secret
    RegisterProductRequest:
      type: object
      properties:
        id:
          type: string
        host_id:
          type: string
        name:
          type: string
        sku:
          type: string
        price:
          type: integer
        is_active:
          type: boolean
        meta:
          type: object
          additionalProperties:
            type: string
        fee_policy_override:
          $ref: '#/components/schemas/FeePolicy'
      required:
        - id
        - host_id
        - name
        - price
    RegisterProviderAccountRequest:
      type: object
      properties:
        host_id:
          type: string
        provider:
          type: string
        env:
          type: string
        credentials_hash:
          type: string
        public_config:
          type: object
          additionalProperties:
            type: string
      required:
        - host_id
        - provider
        - env
        - credentials_hash
    RegisterHostPolicyRequest:
      type: object
      properties:
        host_id:
          type: string
        fee_policy:
          $ref: '#/components/schemas/FeePolicy'
      required:
        - host_id
        - fee_policy
security:
  - apiKey: []
paths:
  /payments:
    post:
      summary: Create payment order (local orchestration path)
      security:
        - apiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentRequest'
      responses:
        '202':
          description: Payment intent created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatePaymentResponse'
  /payments/{reference}:
    get:
      summary: Get payment status
      parameters:
        - in: path
          name: reference
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Payment status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentStatus'
  /webhooks/{provider}:
    post:
      summary: Receive provider webhook
      parameters:
        - in: path
          name: provider
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookPayload'
      responses:
        '202':
          description: Webhook accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookResponse'
  /register/host:
    post:
      summary: Register host profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterHostRequest'
      responses:
        '201':
          description: Host registered
  /register/product:
    post:
      summary: Register host product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterProductRequest'
      responses:
        '201':
          description: Product registered
  /register/provider-account:
    post:
      summary: Register gateway credential mapping per host
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterProviderAccountRequest'
      responses:
        '201':
          description: Provider account registered
  /register/host-policy:
    post:
      summary: Register host fee policy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterHostPolicyRequest'
      responses:
        '201':
          description: Host policy registered
  /host/{host_id}/payments:
    post:
      summary: Create payment via host-scoped proxy to rute-bayar contract
      parameters:
        - in: path
          name: host_id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentRequest'
      responses:
        '200':
          description: Payment intent created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatePaymentResponse'
  /host/{host_id}/payments/{reference}:
    get:
      summary: Get payment status via host-scoped proxy
      parameters:
        - in: path
          name: host_id
          required: true
          schema:
            type: string
        - in: path
          name: reference
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Payment status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentStatus'